pyproject.toml
[tool.poetry]
name = "djangocms-boilerplate"
version = "0.1.0"
description = "DjangoCMS Boilerplate"
authors = ["Dominique Rehborn <dominique@rehborn.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.7"
django-cms = "^3.11.4"
django-sekizai = "^4.1.0"
django-filer = "^3.0.6"
easy-thumbnails = "^2.8.5"
django-mptt = "^0.14.0"
djangocms-text-ckeditor = "^5.1.3"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
django-storages = "^1.14"
sentry-sdk = {extras = ["django"], version = "^1.30.0"}
django-modeltranslation = "^0.18.11"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
psycopg2 = "^2.9.7"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
